+++ drake/systems/plants/test/urdfKinTest.cpp
@@ -33,7 +33,7 @@ int main(int argc, char* argv[]) {
   Eigen::Vector3d zero = Eigen::Vector3d::Zero();
   Eigen::Matrix<double, 6, 1> pt;
 
-  for (i = 0; i < model->bodies.size(); i++) {
+  for (i = 0; i < static_cast<int>(model->bodies.size()); i++) {
     //    model->forwardKin(i, zero, 1, pt);
     auto pt = model->transformPoints(cache, zero, i, 0);
     auto rpy = model->relativeRollPitchYaw(cache, i, 0);
